requirements.txt
gen3_tracker>=0.0.2rc1
click>=7.1.2,<=8.0.4
pandas
numpy
openpyxl
PyYAML
orjson